requirements.txt
-i https://pypi.org/simple
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
bcrypt==3.2.0
bleach==3.2.1
branca==0.4.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
docutils==0.16
folium==0.11.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geoip2==4.1.0
geopy==2.0.0
idna==2.10
ip2geotools==0.1.5
ip2location==8.5.1
isort==5.6.4
jeepney==0.6.0 ; sys_platform == 'linux'
jinja2==2.11.2
keyring==21.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
markupsafe==1.1.1
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.4
packaging==20.8
pandas==1.2.0
paramiko==2.7.2
pip-review==1.1.0
pkginfo==1.6.1
pycodestyle==2.6.0
pycparser==2.20
pygments==2.7.3
pylint==2.6.0
pymysql==0.10.1
pynacl==1.4.0
pyparsing==2.4.7
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2020.5
ratelim==0.1.6
readme-renderer==28.0
requests-toolbelt==0.9.1
requests==2.25.1
rfc3986==1.4.0
secretstorage==3.3.0 ; sys_platform == 'linux'
selenium==3.141.0
six==1.15.0
sshtunnel==0.3.1
toml==0.10.2
tqdm==4.55.0
twine==3.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing==3.7.4.3
urllib3==1.26.2
webencodings==0.5.1
wrapt==1.12.1
yarl==1.6.3
